Integrated File Uploader code residing in /code folder into syllabi-app/frontend folder
