--- code/uploader/postcss.config.mjs
@@ -1,7 +1,0 @@
-const config = {
-  plugins: {
-    "@tailwindcss/postcss": {},
-  },
-};
-
-export default config;
