+++ syllabi-app/frontend/.env.local.EXAMPLE
@@ -1,0 +1,14 @@
+# File: .env.local.EXAMPLE
+# File Upload Configurations
+UPLOAD_DIR=uploads
+MAX_FILE_SIZE=10485760
+MAX_FILES_PER_REQUEST=10
+ALLOWED_FILE_TYPES=application/pdf
+FILE_RETENTION_HOURS=24
+
+NODE_ENV=development
+PORT=3000
+
+# Instructions:
+# 1. Copy this file to .env.local
+# Note: For now, these values are the ones we will be using for the web app
