+++ syllabi-app/frontend/.gitignore
@@ -1,0 +1,9 @@
+# File: .gitignore
+.env
+.env*.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# Commit template
+!.env.local.EXAMPLE
