+++ syllabi-app/frontend/lib/config.ts
