+++ syllabi-app/frontend/lib/errorHandler.ts
