+++ syllabi-app/frontend/lib/fileValidation.ts
