+++ syllabi-app/frontend/lib/logger.ts
