+++ syllabi-app/frontend/src/app/api/cleanup/route.ts
