+++ syllabi-app/frontend/src/app/api/files/[filename]/route.ts
