+++ syllabi-app/frontend/src/app/api/health/route.ts
