+++ syllabi-app/frontend/src/app/api/upload/route.ts
