+++ syllabi-app/frontend/types/events.ts
