+++ syllabi-app/frontend/types/upload.ts
